pyproject.toml
[project]
name = "discord-comfyui"
version = "0.1.0"
description = "A Discord bot for interacting with ComfyUI"
authors = [
    {name = "Joseph Clark", email = "joseph@jclark.codes"},
]
dependencies = [
    "websockets>=12.0",
    "httpx>=0.28.0",
    "discord.py>=2.3.2",
    "PyYAML>=6.0",
    "wat>=0.5.0"
]
requires-python = ">=3.10"

[project.scripts]
discord-comfyui = "discord_comfyui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_comfyui"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]
